dev-agent: restart even if no pid

+++ dev-agent/runner.js
@@ -66,7 +66,8 @@ const start = ({ wkdir, cfg }) => {
     const exitHandler = (code, signal) => {
         console.log('child.on exit', code, signal)
         if (signal !== 'SIGTERM') {
-            runnerOut.emit('message', ({ type: 'exited' }))
+            runnerOut.emit('message', ({ type: 'crash' }))
+            child = null
         } else {
             child.stdout.removeListener('data', stdoutHandler)
             child.stderr.removeListener('data', stderrHandler)
@@ -93,7 +94,17 @@ const start = ({ wkdir, cfg }) => {
 
     runnerIn.on('restart', () => {
         console.log('got restart')
-        process.kill(-child.pid)
+        if (child) {
+            process.kill(-child.pid)
+        } else {
+            child = spawn(...forkArgs)
+            child.on('close', exitHandler)
+
+            child.stdout.on('data', stdoutHandler)
+            child.stderr.on('data', stderrHandler)
+
+            child.on('error', console.error)
+        }
     })
 
     return {
